rename and error display
